Split composition layers to their own files
