+++ examples/composition_layer_quad/game.cpp
@@ -1,6 +1,9 @@
 
 #include "game.h"
 
+#include <vsgvr/app/CompositionLayerProjection.h>
+#include <vsgvr/app/CompositionLayerQuad.h>
+
 #include "../../models/controller/controller.cpp"
 #include "../../models/controller/controller2.cpp"
 #include "../../models/world/world.cpp"
